+++ lib/wisper_next/subscriber/resolve_method.rb
@@ -2,6 +2,8 @@ module WisperNext
   class Subscriber
     class ResolveMethod
       def self.call(name, prefix)
+        name = name.underscore if name.respond_to?(:underscore)
+
         if prefix
           "on_#{name}"
         else