+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added: methods to handle events are underscored when String#underscore is
+  defined, e.g. `def MyEvent` versus `def my_event`.
+
 ## [0.1.0]
 
 - Added: Ability for an object to publish events